[refactor] Cleanup python imports (#2226)
