+++ python/taichi/core/util.py
@@ -208,7 +208,6 @@ else:
         try:
             import_ti_core(tmp_dir)
         except Exception as e:
-            from colorama import Fore, Back, Style
             print_red_bold("Taichi core import failed: ", end='')
             print(e)
             print(
