+++ python/taichi/lang/type_factory_impl.py
@@ -1,0 +1,32 @@
+from taichi.core import util as cutil
+from taichi.lang import impl
+
+
+class TypeFactory:
+    def __init__(self):
+        self.core = cutil.ti_core.get_type_factory_instance()
+
+    def custom_int(self, bits, signed=True, compute_type=None):
+        if compute_type is None:
+            compute_type = impl.get_runtime().default_ip
+        if isinstance(compute_type, cutil.ti_core.DataType):
+            compute_type = compute_type.get_ptr()
+        return self.core.get_custom_int_type(bits, signed, compute_type)
+
+    def custom_float(self,
+                     significand_type,
+                     exponent_type=None,
+                     compute_type=None,
+                     scale=1.0):
+        if compute_type is None:
+            compute_type = impl.get_runtime().default_fp
+        if isinstance(compute_type, cutil.ti_core.DataType):
+            compute_type = compute_type.get_ptr()
+        return self.core.get_custom_float_type(significand_type,
+                                               exponent_type,
+                                               compute_type,
+                                               scale=scale)
+
+
+# Unstable API
+type_factory = TypeFactory()
