+++ python/taichi/misc/task.py
@@ -1,7 +1,36 @@
-from taichi.core import unit
+from taichi.core import util
+from taichi.misc.util import config_from_dict
 
 
-@unit('task')
+def _unit(unit_name):
+    def decorator(target_class):
+        if target_class.__init__ != object.__init__:
+            original_init = target_class.__init__
+        else:
+
+            def dummy_init(*args, **kwargs):
+                pass
+
+            original_init = dummy_init
+
+        def new_init(self, name, *args, **kwargs):
+            self.c = getattr(util.ti_core, 'create_' + unit_name)(name)
+            self.c.initialize(config_from_dict(kwargs))
+            original_init(self, *args, **kwargs)
+
+        target_class.__init__ = new_init
+
+        def new_getattr_(self, key):
+            return self.c.__getattribute__(key)
+
+        target_class.__getattr__ = new_getattr_
+
+        return target_class
+
+    return decorator
+
+
+@_unit('task')
 class Task:
     def run(self, *args):
         return self.c.run(args)
