+++ python/taichi/misc/util.py
@@ -100,18 +100,6 @@ class Tee():
         self.file.write(data)
 
 
-def get_file_name(asc=0):
-    return inspect.stack()[1 + asc][1]
-
-
-def get_function_name(asc=0):
-    return inspect.stack()[1 + asc][3]
-
-
-def get_line_number(asc=0):
-    return inspect.stack()[1 + asc][2]
-
-
 # The builtin `warnings` module is unreliable since it may be suppressed
 # by other packages such as IPython.
 def warning(msg, type=UserWarning, stacklevel=1):
@@ -161,61 +149,12 @@ def obsolete(old, new):
 
 
 def get_traceback(stacklevel=1):
-    import traceback
     s = traceback.extract_stack()[:-1 - stacklevel]
     return ''.join(traceback.format_list(s))
 
 
-def get_logging(name):
-    def logger(msg, *args, **kwargs):
-        # Python inspection takes time (~0.1ms) so avoid it as much as possible
-        if taichi.ti_core.logging_effective(name):
-            import os
-            msg_formatted = msg.format(*args, **kwargs)
-            func = getattr(taichi.ti_core, name)
-            frame = inspect.currentframe().f_back
-            file_name, lineno, func_name, _, _ = inspect.getframeinfo(frame)
-            file_name = os.path.basename(file_name)
-            msg = f'[{file_name}:{func_name}@{lineno}] {msg_formatted}'
-            func(msg)
-
-    return logger
-
-
-DEBUG = 'debug'
-TRACE = 'trace'
-INFO = 'info'
-WARN = 'warn'
-ERROR = 'error'
-CRITICAL = 'critical'
-
-debug = get_logging(DEBUG)
-trace = get_logging(TRACE)
-info = get_logging(INFO)
-warn = get_logging(WARN)
-error = get_logging(ERROR)
-critical = get_logging(CRITICAL)
-
-
-def redirect_print_to_log():
-    class Logger:
-        def write(self, msg):
-            taichi.core.info('[{}:{}@{}] {}'.format(get_file_name(1),
-                                                    get_function_name(1),
-                                                    get_line_number(1), msg))
-
-        def flush(self):
-            taichi.core.flush_log()
-
-    sys.stdout = Logger()
-
-
 def duplicate_stdout_to_file(fn):
     taichi.ti_core.duplicate_stdout_to_file(fn)
-
-
-def set_logging_level(level):
-    taichi.ti_core.set_logging_level(level)
 
 
 def set_gdb_trigger(on=True):
@@ -300,15 +239,4 @@ __all__ = [
     'set_gdb_trigger',
     'print_profile_info',
     'clear_profile_info',
-    'set_logging_level',
-    'info',
-    'warn',
-    'error',
-    'debug',
-    'trace',
-    'INFO',
-    'WARN',
-    'ERROR',
-    'DEBUG',
-    'TRACE',
 ]
